+++ libcaf_core/caf/flow/generation.test.cpp
@@ -317,6 +317,48 @@ SCENARIO("asynchronous buffers can generate flow items") {
   }
 }
 
+// GH-2395 had this sequence of events on from_resource_sub leading to a crash:
+// 1. `on_producer_wakeup` gets called -> schedules action A
+// 2. `request(n)` gets called (demand was 0) -> schedules action B
+// 3. action A runs -> sets `running_` back to false
+// 4. action B runs
+// 5. the observer cancels the subscription mid-batch
+// 6. do_run() still runs, yet cancel() gets called with `running_` is false
+TEST("GH-2395 regression") {
+  // Get our pull and push resources.
+  auto [pull, push] = async::make_spsc_buffer_resource<int>(50, 10);
+  // Something to write to the buffer.
+  auto producer = async::make_blocking_producer(std::move(push));
+  require(producer.has_value());
+  // An observer that will cancel on the first `on_next` call.
+  auto out = make_canceling_observer<int>(true, false);
+  // Our `from_resource_sub` for the regression test.
+  auto buf = pull.try_open();
+  using buffer_type = async::spsc_buffer<int>;
+  using impl_t = caf::flow::op::from_resource_sub<buffer_type>;
+  auto ptr = coordinator()->add_child(std::in_place_type<impl_t>, buf,
+                                      out->as_observer());
+  buf->set_consumer(ptr);
+  out->on_subscribe(subscription{ptr});
+  // Drain any scheduled action.
+  run_flows();
+  // Step 1: trigger `on_producer_wakeup` to schedule action A.
+  ptr->on_producer_wakeup();
+  // Put an action on top that will push to the buffer after action A runs.
+  coordinator()->schedule_fn([this, &producer] {
+    // Push items to the buffer so action B will call `on_next` on the observer.
+    for (int i = 0; i < 10; ++i) {
+      if (!producer->push(i)) {
+        fail("push failed");
+      }
+    }
+  });
+  // Step 2: call `request(n)` to schedule action B.
+  out->sub.request(5);
+  // Prior to the fix for GH-2395, this crashed in `observer<T>::on_next`.
+  run_flows();
+}
+
 namespace {
 
 // Generates 7 integers and then calls on_complete.